+++ cli/js/99_main.js
@@ -1,0 +1,2 @@
+import "ext:cli/40_testing.js";
+import "ext:cli/runtime/js/99_main.js";
