+++ ext/web/09_file.js
@@ -14,6 +14,7 @@ const core = globalThis.Deno.core;
 const ops = core.ops;
 import * as webidl from "ext:deno_webidl/00_webidl.js";
 import { ReadableStream } from "ext:deno_web/06_streams.js";
+import { URL } from "ext:deno_url/00_url.js";
 const primordials = globalThis.__bootstrap.primordials;
 const {
   ArrayBufferPrototype,
@@ -652,6 +653,33 @@ function blobFromObjectUrl(url) {
   blob[_parts] = parts;
   return blob;
 }
+
+/**
+ * @param {Blob} blob
+ * @returns {string}
+ */
+function createObjectURL(blob) {
+  const prefix = "Failed to execute 'createObjectURL' on 'URL'";
+  webidl.requiredArguments(arguments.length, 1, prefix);
+  blob = webidl.converters["Blob"](blob, prefix, "Argument 1");
+
+  return ops.op_blob_create_object_url(blob.type, getParts(blob));
+}
+
+/**
+ * @param {string} url
+ * @returns {void}
+ */
+function revokeObjectURL(url) {
+  const prefix = "Failed to execute 'revokeObjectURL' on 'URL'";
+  webidl.requiredArguments(arguments.length, 1, prefix);
+  url = webidl.converters["DOMString"](url, prefix, "Argument 1");
+
+  ops.op_blob_revoke_object_url(url);
+}
+
+URL.createObjectURL = createObjectURL;
+URL.revokeObjectURL = revokeObjectURL;
 
 export {
   Blob,
