+++ ext/web/lib.rs
@@ -103,7 +103,6 @@ deno_core::extension!(deno_web,
     "08_text_encoding.js",
     "09_file.js",
     "10_filereader.js",
-    "11_blob_url.js",
     "12_location.js",
     "13_message_port.js",
     "14_compression.js",
